Create Code.cpp
This code takes an adjacency matrix as input and max no of chairs and tables used in the wedding. It then gives an arrangement where all the guests who do not want to sit together are seated separately, thereby reducing chances of unwanted fights during the function :)

It uses the concept of graph coloring which is implemented using backtracking. The adjacency matrix is as follows: If in the matrix A , guest i does not want to sit with guest j (i,j) are distinct then Aij = Aji = 1.